Fix RuntimeError: Input type (double) and bias type (float) mismatch in VAE encoding
- Add .float() conversion in encode_pixels method to ensure torch.float32 dtype
- Resolves training failure when pixel data is loaded as torch.float64
- Fixes issue in modules/sd_model_diffusers.py line 78"

+++ modules/sd_model_diffusers.py
@@ -75,7 +75,7 @@ class StableDiffusionModel(pl.LightningModule):
         latents = []
         for i in range(0, feed_pixel_values.shape[0], self.vae_encode_bsz):
             with torch.autocast("cuda", enabled=False):
-                lat = self.vae.encode(feed_pixel_values[i : i + self.vae_encode_bsz]).latent_dist.sample()
+                lat = self.vae.encode(feed_pixel_values.float()[i : i + self.vae_encode_bsz]).latent_dist.sample()
             latents.append(lat)
         latents = torch.cat(latents, dim=0)
         latents = latents * self.vae.config.scaling_factor
